Price list cleanup.
Added entries for a few requests: horses, crossbow, boats and other transport.

Revised all price entries to be roughly in line with other editions. All prices are now somewhere in the neighborhood of the same item in at least one other edition, and as close to all editions as I could manage.

Gold has been renamed Coin. Prices are now in coins.

Closes issue #1.