Made adding spells to the Spellbook not Wizard-specific.
Characters that choose a multiclass advanced move and take the Wizard's
Spellbook move get spells whenever they level up.  Changed the text of the
Level Up move to say you add a spell "if you have the Spellbook move (such
as the Wizard)" rather than "if you are the Wizard".