Reorganize files for more hackability
The old structure and file naming was based around one canonical layout. The chapter number was included in the filename so that the alphabetical sort order was also the layout order.

Then two things happened:
-We added the Druid smack dab in the middle of the classes
-We got back editor feedback from Jay that certain chapters needed to be reordered

Both of those pointed out how brittle this system was. Adding the Druid file meant renaming all the subsequent XML files AND relinking all the layout files to deal with the changed filenames. That just isn't practical.

This also pointed out that we could make the files more hackable by not baking order into the files themselves. As long as the files have an order built into their names we're tacitly discouraging the hacking of the order of chapters. Once we realized we (DWHQ) wanted to reorder chapters it seemed like a good idea to make it easy for everyone else to.

This change does regress usability of the plain files a bit though. Before an interested reader could read the XML files in the same order as the layout had them just by going in order. We plan to return that capability in another form in a subsequent change.

Have fun hacking the order of the text!