Updated thief's starting pack of arrows.
Lucky Thief! He starts with 5 ammo in his quiver instead of the silly
Bard and Ranger and their paltry 3 ammo. ;)

I don't know if this is a design intention, but it seems to break the
spirit. If they need more starting ammo, give the Thief two bundles of
arrows, give them that.